Simplify list updates

+++ src/candidate_lists/structs/candidate.rs
@@ -5,6 +5,6 @@ use crate::{candidate_lists::CandidateListId, persons::Person};
 #[derive(Debug, Clone, Serialize)]
 pub struct Candidate {
     pub list_id: CandidateListId,
-    pub position: i32,
+    pub position: usize,
     pub person: Person,
 }
